Major updates to performance, treating skills as a calculated field

+++ src/actions/mutations/details/delete-detail.ts
@@ -1,26 +1,26 @@
 "use server";
 import { getAuthedUser } from "~/application/auth/server";
-import { isPrismaDoesNotExistError, isPrismaInvalidIdError, prisma } from "~/prisma/client";
-import { type Detail } from "~/prisma/model";
+import { prisma } from "~/prisma/client";
+import { calculateSkillsExperience } from "~/prisma/model";
 import { ApiClientGlobalError } from "~/api";
 
 export const deleteDetail = async (id: string) => {
-  await getAuthedUser();
+  const { user } = await getAuthedUser();
 
   return await prisma.$transaction(async tx => {
-    let detail: Detail;
-    try {
-      detail = await tx.detail.findUniqueOrThrow({
-        where: { id },
-      });
-    } catch (e) {
-      if (isPrismaDoesNotExistError(e) || isPrismaInvalidIdError(e)) {
-        throw ApiClientGlobalError.NotFound();
-      }
-      throw e;
+    const detail = await tx.detail.findUnique({
+      where: { id },
+      include: { skills: true, nestedDetails: { include: { skills: true } } },
+    });
+    if (!detail) {
+      throw ApiClientGlobalError.NotFound();
     }
-
+    const sks = [
+      ...detail.skills.map(s => s.id),
+      ...detail.nestedDetails.flatMap(d => d.skills.map(s => s.id)),
+    ];
     await tx.nestedDetail.deleteMany({ where: { detailId: detail.id } });
     await tx.detail.delete({ where: { id: detail.id } });
+    await calculateSkillsExperience(tx, sks, { user });
   });
 };
